More improvements to export statements
