+++ src/extension/executors/task.ts
@@ -1,5 +1,4 @@
 import path from 'node:path'
-import cp from 'node:child_process'
 
 import {
   Task, TextDocument, NotebookCellExecution, TaskScope, tasks,
@@ -8,16 +7,15 @@ import {
 } from 'vscode'
 
 // import { ExperimentalTerminal } from "../terminal"
-import { populateEnvVar, getExecutionProperty, getCmdShellSeq, } from '../utils'
-import { ENV_STORE, PLATFORM_OS } from '../constants'
+import { getExecutionProperty, getCmdShellSeq } from '../utils'
+import { PLATFORM_OS, ENV_STORE } from '../constants'
 import type { Kernel } from '../kernel'
 
+import { retrieveShellCommand } from './utils'
 import { sh as inlineSh } from './shell'
 
 const BACKGROUND_TASK_HIDE_TIMEOUT = 2000
 const LABEL_LIMIT = 15
-const EXPORT_EXTRACT_REGEX = /(\n*)export \w+=(("(.|\n)+(?="))|(.+(?=\n)))/gi
-const EXPORT_REGEX = /(\n*)export \w+=/gi
 
 export function closeTerminalByEnvID (id: string) {
   const terminal = window.terminals.find((t) => (t.creationOptions as TerminalOptions).env?.RUNME_ID === id)
@@ -32,87 +30,12 @@ async function taskExecutor(
   doc: TextDocument
 ): Promise<boolean> {
   const cwd = path.dirname(doc.uri.fsPath)
-  let cellText = doc.getText()
-
-  /**
-   * find export commands
-   */
-  const rawText = doc.getText()
-  const isSingleLine = (rawText.match(EXPORT_REGEX) || []).length <= 1
-  const lines: string[] = isSingleLine ? [rawText] : rawText.split('\n')
-  const exportMatches: string[] = []
-  for (const l of lines) {
-    const code = l.endsWith('\n') ? l : `${l}\n`
-    const exps = (code.match(EXPORT_EXTRACT_REGEX) || [])
-      .map((m) => m.trim())
-    exportMatches.push(...exps)
+  const cellText = await retrieveShellCommand(exec)
+  if (!cellText) {
+    return false
   }
+
   const stateEnv = Object.fromEntries(ENV_STORE)
-  for (const e of exportMatches) {
-    const [key, ph] = e.slice('export '.length).split('=')
-    const hasStringValue = ph.startsWith('"')
-    const placeHolder = hasStringValue ? ph.slice(1) : ph
-
-    if (placeHolder.startsWith('$(') && placeHolder.endsWith(')')) {
-      /**
-       * evaluate expression
-       */
-      const expression = placeHolder.slice(2, -1)
-      const expressionProcess = cp.spawn(expression, {
-        cwd,
-        shell: true
-      })
-      const [isError, data] = await new Promise<[number, string]>((resolve) => {
-        let data = ''
-        expressionProcess.stdout.on('data', (payload) => { data += payload.toString() })
-        expressionProcess.stderr.on('data', (payload) => { data += payload.toString() })
-        expressionProcess.on('close', (code) => {
-          if (code && code > 0) {
-            return resolve([code, data])
-          }
-
-          return resolve([0, data])
-        })
-      })
-
-      if (isError) {
-        window.showErrorMessage(`Failed to evaluate expression "${expression}": ${data}`)
-        return false
-      }
-
-      stateEnv[key] = data
-    } else {
-      /**
-       * ask user for value
-       */
-      stateEnv[key] = populateEnvVar(await window.showInputBox({
-        title: `Set Environment Variable "${key}"`,
-        ignoreFocusOut: true,
-        placeHolder,
-        prompt: 'Your shell script wants to set some environment variables, please enter them here.',
-        ...(hasStringValue ? { value: placeHolder } : {})
-      }) || '', {...process.env, ...stateEnv })
-    }
-
-    /**
-     * we don't want to run these exports anymore as we already stored
-     * them in our extension state
-     */
-    cellText = cellText.replace(
-      /**
-       * In case of `export foo="bar", our match includes the preceeding '"' but not
-       * the ending one. To properly cut out the line from the script we need to add
-       * it back again.
-       */
-      e + (hasStringValue ? '"' : ''),
-      ''
-    )
-
-    /**
-     * persist env variable in memory
-     */
-    ENV_STORE.set(key, stateEnv[key])
-  }
 
   const RUNME_ID = `${doc.fileName}:${exec.cell.index}`
   const env = {
